toolchain_build: Add source packages, incremental builds and LLVM build
This CL introduces the concept of source vs. build targets in the build
scripts. Source targets are not memoized and can be depended on by
build targets (which allows build targets to share source targets and to
depend on multiple sources). It also allows disabling syncing sources
via a flag. Targets can also now have an 'inputs' key which specifies file or
directory paths that get used as inputs in addition to the dependency outputs.

It also introduces the concept of incremental builds (currently enabled
via a flag). Incremental builds do not clobber the working directory before
they run, and Commands on build targets can be marked with the
skip_for_incremental attribute, which will cause them
to be skipped for incremental builds.

These concepts are used in a strightforward way to implement a build
of LLVM and the PNaCl driver scripts.

This is a re-land of r12429 (and r12430) with the additional change that
once.Run creates the output directory if it is absent.

R=bradnelson@google.com
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3719

Review URL: https://codereview.chromium.org/82553011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12441 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 994b16d23d9ca0c3eddad3807fbe52b31ca8517b

+++ build/working_directory.py
@@ -14,30 +14,29 @@ import tempfile
 class FixedWorkingDirectory(object):
   """Working directory manager to be used with 'with'.
 
-  The manager cleanups up the directory initially if it's
-  there, but leaves it in place at the end.
-  Done in this style for drop in compatibility with TemporaryWorkingDirectory
-  below.
+  The context manager does not clean the directory on entry or exit.
 
   Example uses:
-    for WorkingDirectory('mydir') as work_dir:
+    with WorkingDirectory('mydir', clobber=True) as work_dir:
       ...do stuff in work_dir...
   """
 
-  def __init__(self, work_dir=None):
+  def __init__(self, work_dir=None, clobber=False):
     """ Constructor.
 
     Args:
       work_dir: A selected working directory.
     """
     self._work_dir = os.path.abspath(work_dir)
+    self._clobber = clobber
 
   def __enter__(self):
-    if os.path.exists(self._work_dir):
+    if self._clobber and os.path.exists(self._work_dir):
       logging.debug('Removing %s...' % self._work_dir)
       shutil.rmtree(self._work_dir)
-    logging.debug('Creating %s...' % self._work_dir)
-    os.mkdir(self._work_dir)
+    if not os.path.exists(self._work_dir):
+      logging.debug('Creating %s...' % self._work_dir)
+      os.mkdir(self._work_dir)
     return self._work_dir
 
   def __exit__(self, _type, _value, _trackback):
@@ -51,7 +50,7 @@ class TemporaryWorkingDirectory(object):
   up at the end.
 
   Example uses:
-    for TemporaryWorkingDirectory() as work_dir:
+    with TemporaryWorkingDirectory() as work_dir:
       ...do stuff in work_dir...
   """
   def __enter__(self):
